refactor(test): extract shared D1 fixture helpers to helpers/fixtures.ts
setup.ts and global-setup.ts each inlined ~60 lines of identical SQL
helpers and fixture constants. The duplication was a drift hazard — any
change to the fixture seed had to be made in lockstep in both files.

Move stripSqlCommentsAndCollapse, buildContractsInsert, and the seven
FIXTURE_* constants (plus a canonical FIXTURE_STATEMENTS array) into a
new helpers/fixtures.ts module that both files now import. Behaviour is
unchanged: same SQL emitted, same apply order, same INSERT OR IGNORE
semantics.

Test count and outputs unchanged: pnpm --filter @sigma/web test still
runs 335 unit + 34 integration = 369 tests, all green.

Reviewer note: addresses the duplication comment from #177 (ydimitrof
on c733245).
